seperated one-ink fetch into singular state
